Move TransformlistProperty into own file
