+++ modules/base/include/modules/base/processors/transform.h
@@ -30,35 +30,16 @@
 #pragma once
 
 #include <modules/base/basemoduledefine.h>
+#include <modules/base/properties/transformlistproperty.h>
 #include <inviwo/core/common/inviwo.h>
 #include <inviwo/core/datastructures/coordinatetransformer.h>
 #include <inviwo/core/processors/processor.h>
-#include <inviwo/core/properties/compositeproperty.h>
-#include <inviwo/core/properties/listproperty.h>
 #include <inviwo/core/properties/boolproperty.h>
-#include <inviwo/core/properties/ordinalproperty.h>
 #include <inviwo/core/properties/optionproperty.h>
 #include <inviwo/core/ports/datainport.h>
 #include <inviwo/core/ports/dataoutport.h>
 
 namespace inviwo {
-
-/** \docpage{org.inviwo.TransformListProperty, Transform List Property}
- * ![](org.inviwo.TransformListProperty.png?classIdentifier=org.inviwo.TransformListProperty)
- * List of transformations being applied on each other.
- */
-class IVW_MODULE_BASE_API TransformListProperty : public CompositeProperty {
-public:
-    TransformListProperty(const std::string& identifier, const std::string& displayName,
-                          InvalidationLevel invalidationLevel = InvalidationLevel::InvalidOutput,
-                          PropertySemantics semantics = PropertySemantics::Default);
-    ~TransformListProperty() = default;
-
-    const mat4& getMatrix() const;
-
-    ListProperty transforms_;
-    FloatMat4Property result_;
-};
 
 template <typename T>
 class Transform : public Processor {
@@ -119,107 +100,6 @@ struct ProcessorTraits<Transform<Volume>> {
         };
     }
 };
-
-namespace transform {
-
-class IVW_MODULE_BASE_API TransformProperty : public CompositeProperty {
-public:
-    TransformProperty(const std::string& identifier, const std::string& displayName,
-                      InvalidationLevel invalidationLevel = InvalidationLevel::InvalidResources,
-                      PropertySemantics semantics = PropertySemantics::Default);
-    virtual ~TransformProperty() = default;
-
-    virtual mat4 getMatrix() const = 0;
-};
-
-class IVW_MODULE_BASE_API TranslateProperty : public TransformProperty {
-public:
-    virtual std::string getClassIdentifier() const override;
-    static const std::string classIdentifier;
-
-    TranslateProperty(const std::string& identifier, const std::string& displayName,
-                      const vec3& value = vec3{0.0f}, const vec3& minValue = vec3{-1.e6f},
-                      const vec3& maxValue = vec3{1.e6f},
-                      const vec3& increment = Defaultvalues<vec3>::getInc(),
-                      InvalidationLevel invalidationLevel = InvalidationLevel::InvalidResources,
-                      PropertySemantics semantics = PropertySemantics::Default);
-    TranslateProperty(const TranslateProperty& rhs);
-    virtual TranslateProperty* clone() const override;
-    virtual ~TranslateProperty() = default;
-
-    virtual mat4 getMatrix() const override;
-
-    FloatVec3Property translate;
-};
-
-class IVW_MODULE_BASE_API RotateProperty : public TransformProperty {
-public:
-    enum class AngleMeasure { Radians, Degrees };
-
-    virtual std::string getClassIdentifier() const override;
-    static const std::string classIdentifier;
-
-    RotateProperty(const std::string& identifier, const std::string& displayName,
-                   const vec3& axis = vec3{1.0f, 0.0f, 0.0f}, const float angle = 0.0f,
-                   const float minAngle = -glm::pi<float>(),
-                   const float maxAngle = glm::pi<float>(),
-                   const float increment = Defaultvalues<float>::getInc(),
-                   AngleMeasure angleMeasure = AngleMeasure::Radians,
-                   InvalidationLevel invalidationLevel = InvalidationLevel::InvalidResources,
-                   PropertySemantics semantics = PropertySemantics::Default);
-    RotateProperty(const RotateProperty& rhs);
-    virtual RotateProperty* clone() const override;
-    virtual ~RotateProperty() = default;
-
-    virtual mat4 getMatrix() const override;
-
-    TemplateOptionProperty<AngleMeasure> mode;
-    FloatVec3Property axis;
-    FloatProperty angle;
-};
-
-class IVW_MODULE_BASE_API ScaleProperty : public TransformProperty {
-public:
-    virtual std::string getClassIdentifier() const override;
-    static const std::string classIdentifier;
-
-    ScaleProperty(const std::string& identifier, const std::string& displayName,
-                  const vec3& value = vec3{1.0f}, const vec3& minValue = vec3{-1.e3f},
-                  const vec3& maxValue = vec3{1.e3f},
-                  const vec3& increment = Defaultvalues<vec3>::getInc(),
-                  InvalidationLevel invalidationLevel = InvalidationLevel::InvalidResources,
-                  PropertySemantics semantics = PropertySemantics::Default);
-    ScaleProperty(const ScaleProperty& rhs);
-    virtual ScaleProperty* clone() const override;
-    virtual ~ScaleProperty() = default;
-
-    virtual mat4 getMatrix() const override;
-
-    FloatVec3Property scale;
-};
-
-class IVW_MODULE_BASE_API CustomTransformProperty : public TransformProperty {
-public:
-    virtual std::string getClassIdentifier() const override;
-    static const std::string classIdentifier;
-
-    CustomTransformProperty(
-        const std::string& identifier, const std::string& displayName,
-        const mat4& value = mat4{1.0f}, const mat4& minValue = util::filled<mat4>(-1.e6f),
-        const mat4& maxValue = util::filled<mat4>(1.e6f),
-        const mat4& increment = Defaultvalues<mat4>::getInc(),
-        InvalidationLevel invalidationLevel = InvalidationLevel::InvalidResources,
-        PropertySemantics semantics = PropertySemantics::Default);
-    CustomTransformProperty(const CustomTransformProperty& rhs);
-    virtual CustomTransformProperty* clone() const override;
-    virtual ~CustomTransformProperty() = default;
-
-    virtual mat4 getMatrix() const override;
-
-    FloatMat4Property matrix;
-};
-
-}  // namespace transform
 
 template <typename T>
 Transform<T>::Transform()
