+++ modules/base/src/processors/transform.cpp
@@ -28,175 +28,3 @@
  *********************************************************************************/
 
 #include <modules/base/processors/transform.h>
-
-namespace inviwo {
-
-namespace transform {
-
-TransformProperty::TransformProperty(const std::string& identifier, const std::string& displayName,
-                                     InvalidationLevel invalidationLevel,
-                                     PropertySemantics semantics)
-    : CompositeProperty(identifier, displayName, invalidationLevel, semantics) {}
-
-const std::string TranslateProperty::classIdentifier = "org.inviwo.trafo.TranslateProperty";
-std::string TranslateProperty::getClassIdentifier() const { return classIdentifier; }
-
-TranslateProperty::TranslateProperty(const std::string& identifier, const std::string& displayName,
-                                     const vec3& value, const vec3& minValue, const vec3& maxValue,
-                                     const vec3& increment, InvalidationLevel invalidationLevel,
-                                     PropertySemantics semantics)
-    : TransformProperty(identifier, displayName, invalidationLevel, semantics)
-    , translate("translate_", "Translate", value, minValue, maxValue, increment) {
-    translate.setSemantics(PropertySemantics::SpinBox);
-    addProperty(translate);
-}
-
-TranslateProperty::TranslateProperty(const TranslateProperty& rhs)
-    : TransformProperty(rhs), translate(rhs.translate) {
-    addProperty(translate);
-}
-
-TranslateProperty* TranslateProperty::clone() const { return new TranslateProperty(*this); }
-
-mat4 TranslateProperty::getMatrix() const { return glm::translate(*translate); }
-
-const std::string RotateProperty::classIdentifier = "org.inviwo.trafo.RotateProperty";
-std::string RotateProperty::getClassIdentifier() const { return classIdentifier; }
-
-RotateProperty::RotateProperty(const std::string& identifier, const std::string& displayName,
-                               const vec3& axisValue, const float angleValue, const float minAngle,
-                               const float maxAngle, const float increment,
-                               AngleMeasure angleMeasure, InvalidationLevel invalidationLevel,
-                               PropertySemantics semantics)
-    : TransformProperty(identifier, displayName, invalidationLevel, semantics)
-    , mode("mode", "Angle Measure",
-           {{"rad", "Radians", AngleMeasure::Radians}, {"deg", "Degrees", AngleMeasure::Degrees}})
-    , axis("axis", "Axis", glm::normalize(axisValue), vec3{-1.0f}, vec3{1.0f})
-    , angle("angle", "Angle", angleValue, minAngle, maxAngle, increment) {
-    mode.setSelectedValue(angleMeasure);
-    axis.setSemantics(PropertySemantics::SpinBox);
-    addProperties(mode, axis, angle);
-
-    mode.onChange([this]() {
-        if (mode == AngleMeasure::Radians) {
-            angle.set(glm::radians(*angle), glm::radians(angle.getMinValue()),
-                      glm::radians(angle.getMaxValue()), angle.getIncrement());
-        } else {
-            angle.set(glm::degrees(*angle), glm::degrees(angle.getMinValue()),
-                      glm::degrees(angle.getMaxValue()), angle.getIncrement());
-        }
-    });
-}
-
-RotateProperty::RotateProperty(const RotateProperty& rhs)
-    : TransformProperty(rhs), mode(rhs.mode), axis(rhs.axis), angle(rhs.angle) {
-    addProperties(mode, axis, angle);
-
-    mode.onChange([this]() {
-        if (mode == AngleMeasure::Radians) {
-            angle.set(glm::radians(*angle), glm::radians(angle.getMinValue()),
-                      glm::radians(angle.getMaxValue()), angle.getIncrement());
-        } else {
-            angle.set(glm::degrees(*angle), glm::degrees(angle.getMinValue()),
-                      glm::degrees(angle.getMaxValue()), angle.getIncrement());
-        }
-    });
-}
-
-RotateProperty* RotateProperty::clone() const { return new RotateProperty(*this); }
-
-mat4 RotateProperty::getMatrix() const {
-    const float angleRad = (mode == AngleMeasure::Radians ? angle : glm::radians(*angle));
-    return glm::rotate(angleRad, *axis);
-}
-
-const std::string ScaleProperty::classIdentifier = "org.inviwo.trafo.ScaleProperty";
-std::string ScaleProperty::getClassIdentifier() const { return classIdentifier; }
-
-ScaleProperty::ScaleProperty(const std::string& identifier, const std::string& displayName,
-                             const vec3& value, const vec3& minValue, const vec3& maxValue,
-                             const vec3& increment, InvalidationLevel invalidationLevel,
-                             PropertySemantics semantics)
-    : TransformProperty(identifier, displayName, invalidationLevel, semantics)
-    , scale("scale", "Scale", value, minValue, maxValue, increment) {
-    scale.setSemantics(PropertySemantics::SpinBox);
-    addProperty(scale);
-}
-
-ScaleProperty::ScaleProperty(const ScaleProperty& rhs) : TransformProperty(rhs), scale(rhs.scale) {
-    addProperty(scale);
-}
-
-ScaleProperty* ScaleProperty::clone() const { return new ScaleProperty(*this); }
-
-mat4 ScaleProperty::getMatrix() const { return glm::scale(*scale); }
-
-const std::string CustomTransformProperty::classIdentifier =
-    "org.inviwo.trafo.CustomTransformProperty";
-std::string CustomTransformProperty::getClassIdentifier() const { return classIdentifier; }
-
-CustomTransformProperty::CustomTransformProperty(const std::string& identifier,
-                                                 const std::string& displayName, const mat4& value,
-                                                 const mat4& minValue, const mat4& maxValue,
-                                                 const mat4& increment,
-                                                 InvalidationLevel invalidationLevel,
-                                                 PropertySemantics semantics)
-    : TransformProperty(identifier, displayName, invalidationLevel, semantics)
-    , matrix("matrix", "Matrix", value, minValue, maxValue, increment) {
-    matrix.setSemantics(PropertySemantics::SpinBox);
-    addProperty(matrix);
-}
-
-CustomTransformProperty::CustomTransformProperty(const CustomTransformProperty& rhs)
-    : TransformProperty(rhs), matrix(rhs.matrix) {
-    addProperty(matrix);
-}
-
-CustomTransformProperty* CustomTransformProperty::clone() const {
-    return new CustomTransformProperty(*this);
-}
-
-mat4 CustomTransformProperty::getMatrix() const { return matrix; }
-
-}  // namespace transform
-
-TransformListProperty::TransformListProperty(const std::string& identifier,
-                                             const std::string& displayName,
-                                             InvalidationLevel invalidationLevel,
-                                             PropertySemantics semantics)
-    : CompositeProperty(identifier, displayName, invalidationLevel, semantics)
-    , transforms_(
-          "transforms", "Transformations",
-          []() {
-              std::vector<std::unique_ptr<Property>> v;
-              v.emplace_back(
-                  std::make_unique<transform::TranslateProperty>("translation", "Translation"));
-              v.emplace_back(std::make_unique<transform::RotateProperty>("rotation", "Rotation"));
-              v.emplace_back(std::make_unique<transform::ScaleProperty>("scaling", "Scaling"));
-              v.emplace_back(
-                  std::make_unique<transform::CustomTransformProperty>("custom", "Custom Matrix"));
-              return v;
-          }())
-    , result_(
-          "result", "Result", mat4(1.0f),
-          {util::filled<mat4>(std::numeric_limits<float>::lowest()), ConstraintBehavior::Ignore},
-          {util::filled<mat4>(std::numeric_limits<float>::max()), ConstraintBehavior::Ignore},
-          util::filled<mat4>(0.001f), InvalidationLevel::Valid) {
-    result_.setSemantics(PropertySemantics::Text);
-    result_.setReadOnly(true);
-    addProperties(transforms_, result_);
-
-    transforms_.onChange([this]() {
-        mat4 total{1.0f};
-        for (auto p : transforms_) {
-            if (auto prop = dynamic_cast<transform::TransformProperty*>(p)) {
-                total = prop->getMatrix() * total;
-            }
-        }
-        result_.set(total);
-    });
-}
-
-const mat4& TransformListProperty::getMatrix() const { return result_.get(); }
-
-}  // namespace inviwo
