+++ modules/meshrenderinggl/include/modules/meshrenderinggl/processors/transformrasterization.h
@@ -31,7 +31,7 @@
 
 #include <modules/meshrenderinggl/datastructures/rasterization.h>
 #include <modules/meshrenderinggl/ports/rasterizationport.h>
-#include <modules/base/processors/transform.h>
+#include <modules/base/properties/transformlistproperty.h>
 
 namespace inviwo {
